factored out the ROS stuff into new file
